LayoutTests/imported/blink/fast/events/event-isolated-world-clone-expected.txt

CONSOLE MESSAGE: line 10: CustomEvent received in main world
CONSOLE MESSAGE: line 11: detail was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: line 3: CustomEvent received in isolated world
CONSOLE MESSAGE: line 4: detail was {"foo":5,"bar":"hello","targetWorld":"isolated"}
CONSOLE MESSAGE: line 10: MessageEvent received in main world
CONSOLE MESSAGE: line 11: data was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: line 3: MessageEvent received in isolated world
CONSOLE MESSAGE: line 4: data was {"foo":5,"bar":"hello","targetWorld":"isolated"}
CONSOLE MESSAGE: line 10: PopStateEvent received in main world
CONSOLE MESSAGE: line 11: state was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: line 3: PopStateEvent received in isolated world
CONSOLE MESSAGE: line 4: state was {"foo":5,"bar":"hello","targetWorld":"isolated"}
Tests that properties of various events are cloned between isolated worlds.
